refactor: OpenAPI 클래스 추상화 및 DTO 어노테이션 수정
