+++ src/main/java/parkSystem/park/ParkApplication.java
@@ -1,9 +1,7 @@
 package parkSystem.park;
 
-import lombok.RequiredArgsConstructor;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import parkSystem.park.park.service.ParkOpenApiService;
 
 @SpringBootApplication
 public class ParkApplication {
